[BUGFIX] Elements are always inserted at top
When an element should be inserted after a given element, the sorting
must to be changed accordingly. This was not done until now, but is
added with this commit.

The tests did not uncover this bug because they executed the operation
in a wrong way: with the related content's uid (negated) as the copy
target instead of the page where to insert. Therefore DataHandler already
did the resorting, but as this is an artificial scenario and the real
operation in the page module uses the page as the copy target, we have
to introduce the manual reordering in ContentService.